# docs/_static/files/swagger.json
{
    "swagger": "2.0",
    "basePath": "\/api",
    "paths": {
...
        "\/sighting": {
            "post": {
                "responses": {
                    "201": {
                        "description": "Success.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#\/definitions\/SightingsList"
                            }
                        }
                    },
                    "400": {
                        "description": "Incorrect payload to instantiate a sighting."
                    },
                    "409": {
                        "description": "Duplicate sighting."
                    }
                },
                "summary": "Create a new sighting",
                "description": "Create a new sighting.",
                "operationId": "post_sightings_list",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#\/definitions\/Sighting"
                        }
                    },
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "tags": [
                    "sighting"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#\/definitions\/SightingsList"
                            }
                        }
                    }
                },
                "summary": "List all sightings",
                "operationId": "list_sightings",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "Number of the page.",
                        "default": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "type": "integer",
                        "description": "Maximum number of elements to return.",
                        "default": 1000
                    },
                    {
                        "name": "uuid",
                        "in": "query",
                        "type": "string",
                        "description": "UUID of the sighting."
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "type": "string",
                        "description": "Type of sightings:\n     - Seen: The vulnerability was mentioned, discussed, or seen somewhere by the user.\n     - Exploited: This vulnerability was exploited and seen by the user reporting the sighting.\n     - Not exploited: This vulnerability was not exploited or seen by the user reporting the sighting.\n     - Confirmed: The vulnerability is confirmed from an analyst perspective.\n     - Not confirmed: The user expresses doubt about the veracity of the vulnerability.\n     - Patched: This vulnerability was successfully patched by the user reporting the sighting.\n     - Not patched: This vulnerability was not successfully patched by the user reporting the sighting.",
                        "enum": [
                            "seen",
                            "exploited",
                            "not-exploited",
                            "confirmed",
                            "not-confirmed",
                            "patched",
                            "not-patched"
                        ]
                    },
                    {
                        "name": "vuln_id",
                        "in": "query",
                        "type": "string",
                        "description": "Vulnerability related to the sighting."
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "type": "string",
                        "description": "Author of the sighting (login)."
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "type": "string",
                        "format": "date",
                        "description": "The date of the sightings must be bigger or equal than this value. Format: YYYY-MM-DD"
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "type": "string",
                        "format": "date",
                        "description": "The date of the sightings must be smaller or equal than this value. Format: YYYY-MM-DD"
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "type": "string",
                        "description": "Source of the sighting."
                    },
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "tags": [
                    "sighting"
                ]
            }
        },
        "\/sighting\/": {
            "post": {
                "responses": {
                    "201": {
                        "description": "Success.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#\/definitions\/SightingsList"
                            }
                        }
                    },
                    "400": {
                        "description": "Incorrect payload to instantiate a sighting."
                    },
                    "409": {
                        "description": "Duplicate sighting."
                    }
                },
                "summary": "Create a new sighting",
                "description": "Create a new sighting.",
                "operationId": "post_sightings_list",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#\/definitions\/Sighting"
                        }
                    },
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "tags": [
                    "sighting"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#\/definitions\/SightingsList"
                            }
                        }
                    }
                },
                "summary": "List all sightings",
                "operationId": "list_sightings",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "Number of the page.",
                        "default": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "type": "integer",
                        "description": "Maximum number of elements to return.",
                        "default": 1000
                    },
                    {
                        "name": "uuid",
                        "in": "query",
                        "type": "string",
                        "description": "UUID of the sighting."
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "type": "string",
                        "description": "Type of sightings:\n     - Seen: The vulnerability was mentioned, discussed, or seen somewhere by the user.\n     - Exploited: This vulnerability was exploited and seen by the user reporting the sighting.\n     - Not exploited: This vulnerability was not exploited or seen by the user reporting the sighting.\n     - Confirmed: The vulnerability is confirmed from an analyst perspective.\n     - Not confirmed: The user expresses doubt about the veracity of the vulnerability.\n     - Patched: This vulnerability was successfully patched by the user reporting the sighting.\n     - Not patched: This vulnerability was not successfully patched by the user reporting the sighting.",
                        "enum": [
                            "seen",
                            "exploited",
                            "not-exploited",
                            "confirmed",
                            "not-confirmed",
                            "patched",
                            "not-patched"
                        ]
                    },
                    {
                        "name": "vuln_id",
                        "in": "query",
                        "type": "string",
                        "description": "Vulnerability related to the sighting."
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "type": "string",
                        "description": "Author of the sighting (login)."
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "type": "string",
                        "format": "date",
                        "description": "The date of the sightings must be bigger or equal than this value. Format: YYYY-MM-DD"
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "type": "string",
                        "format": "date",
                        "description": "The date of the sightings must be smaller or equal than this value. Format: YYYY-MM-DD"
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "type": "string",
                        "description": "Source of the sighting."
                    },
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "tags": [
                    "sighting"
                ]
            }
        },
        "\/sighting\/{sighting_uuid}": {
            "parameters": [
                {
                    "name": "sighting_uuid",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success.",
                        "schema": {
                            "$ref": "#\/definitions\/Sighting"
                        }
                    },
                    "404": {
                        "description": "Sighting not found."
                    }
                },
                "description": "Get a sighting with its UUID.",
                "operationId": "get_sighting_item",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "tags": [
                    "sighting"
                ]
            }
        },
...
        "\/user\/": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#\/definitions\/UserFull"
                        }
                    },
                    "201": {
                        "description": "Success."
                    }
                },
                "summary": "Create a non-admin user",
                "description": "Create a non-admin user.\nThe user will be active but must have to confirm the account with the email sent by the instance.",
                "operationId": "post_users_list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "Number of the page.",
                        "default": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "type": "integer",
                        "description": "Maximum number of elements to return.",
                        "default": 10
                    },
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "login": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "organisation": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "tags": [
                    "user"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#\/definitions\/UsersList"
                            }
                        }
                    },
                    "403": {
                        "description": "Admin permission required."
                    }
                },
                "summary": "List all users",
                "description": "Only available to administrators.",
                "operationId": "list_users",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "Number of the page.",
                        "default": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "type": "integer",
                        "description": "Maximum number of elements to return.",
                        "default": 10
                    },
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
...
        "\/vulnerability\/": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success."
                    },
                    "400": {
                        "description": "JSON validation failed."
                    },
                    "403": {
                        "description": "Reporter permission required."
                    },
                    "422": {
                        "description": "Not possible to edit a vulnerability from the requested source."
                    }
                },
                "summary": "Endpoint for creating and editing vulnerabilities in the local source",
                "description": "Create a vulnerability with the CVE version 5 format.",
                "operationId": "post_vulnerabilities_list",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "tags": [
                    "vulnerability"
                ]
            }
        },
...
    },
    "info": {
        "title": "Vulnerability-Lookup API",
        "version": "2.10.0",
        "description": "<a href='https:\/\/github.com\/vulnerability-lookup\/vulnerability-lookup' rel='noreferrer' target='_blank'><img src='https:\/\/vulnerability.circl.lu\/static\/img\/VL-hori-coul.png' width='500px' \/><\/a><br \/>API to query <a href='https:\/\/github.com\/vulnerability-lookup\/vulnerability-lookup' rel='noreferrer' target='_blank'>Vulnerability-Lookup<\/a>. A project from <a href='https:\/\/www.circl.lu' target='_blank'>CIRCL<\/a>.<br \/><br \/>Back to the <a href='\/'>main page<\/a><br \/><br \/><a href='https:\/\/www.vulnerability-lookup.org\/documentation' target='_blank'>Official documentation<\/a> of the project.",
        "license": {
            "name": "GNU Affero General Public License version 3",
            "url": "https:\/\/www.gnu.org\/licenses\/agpl-3.0.html"
        }
    },
    "produces": [
        "application\/json"
    ],
    "consumes": [
        "application\/json"
    ],
    "securityDefinitions": {
        "apikey": {
            "type": "apiKey",
            "in": "header",
            "name": "X-API-KEY"
        }
    },
    "security": [
        {
            "apikey": []
        }
    ],
...
}